+++ examples/feature_demo/pbr.py
@@ -35,9 +35,6 @@ import imageio.v3 as iio
 from wgpu.gui.auto import WgpuCanvas, run
 import pygfx as gfx
 
-from  pygfx.utils.gltf_loader import GLTF
-
-
 # Init
 canvas = WgpuCanvas(size=(640, 480), title="gfx_pbr")
 renderer = gfx.renderers.WgpuRenderer(canvas)
